Updated get id function
